+++ Transit/Transit/Models/NavigationDestination.swift
@@ -6,4 +6,5 @@ enum NavigationDestination: Hashable {
     case projectEdit(Project)
     case milestoneEdit(project: Project, milestone: Milestone?)
     case report
+    case acknowledgments
 }