+++ Transit/Transit/TransitApp.swift
@@ -112,6 +112,8 @@ struct TransitApp: App {
                             MilestoneEditView(project: project, milestone: milestone)
                         case .report:
                             ReportView()
+                        case .acknowledgments:
+                            AcknowledgmentsView()
                         }
                     }
             }