+++ Transit/Transit/Views/Settings/AcknowledgmentsView.swift
@@ -1,0 +1,8 @@
+import SwiftUI
+
+struct AcknowledgmentsView: View {
+    var body: some View {
+        Text("Acknowledgments")
+            .navigationTitle("Acknowledgments")
+    }
+}